+++ lib/rpcs/results.rs
@@ -9,6 +9,7 @@ pub use super::v1_5_0::get_chainspec::{ChainspecRawBytes, GetChainspecResult};
 pub use super::v1_5_0::get_deploy::GetDeployResult;
 pub use super::v1_5_0::get_dictionary_item::GetDictionaryItemResult;
 pub use super::v1_5_0::get_era_info::{EraSummary, GetEraInfoResult};
+pub use super::v1_5_0::get_era_summary::GetEraSummaryResult;
 pub use super::v1_5_0::get_node_status::{
     ActivationPoint, AvailableBlockRange, GetNodeStatusResult, MinimalBlockInfo, NextUpgrade,
     ReactorState,