+++ src/get_era_summary.rs
@@ -7,12 +7,8 @@ use casper_client::cli::CliError;
 
 use crate::{command::ClientCommand, common, Success};
 
-/// Legacy name of command.
-const COMMAND_ALIAS: &str = "get-era-info-by-switch-block";
+pub struct GetEraSummary;
 
-pub struct GetEraInfo;
-
-/// This struct defines the order in which the args are shown for this subcommand's help message.
 enum DisplayOrder {
     Verbose,
     NodeAddress,
@@ -21,14 +17,12 @@ enum DisplayOrder {
 }
 
 #[async_trait]
-impl ClientCommand for GetEraInfo {
-    const NAME: &'static str = "get-era-info";
-    const ABOUT: &'static str = "Retrieve era information at a given switch block";
+impl ClientCommand for GetEraSummary {
+    const NAME: &'static str = "get-era-summary";
+    const ABOUT: &'static str = "Retrieve era information at a given block";
 
     fn build(display_order: usize) -> Command {
         Command::new(Self::NAME)
-            .hide(true)
-            .alias(COMMAND_ALIAS)
             .about(Self::ABOUT)
             .display_order(display_order)
             .arg(common::verbose::arg(DisplayOrder::Verbose as usize))
@@ -48,8 +42,7 @@ impl ClientCommand for GetEraInfo {
         let verbosity_level = common::verbose::get(matches);
         let maybe_block_id = common::block_identifier::get(matches);
 
-        #[allow(deprecated)]
-        casper_client::cli::get_era_info(
+        casper_client::cli::get_era_summary(
             maybe_rpc_id,
             node_address,
             verbosity_level,