+++ CHANGELOG.md
@@ -14,7 +14,7 @@ All notable changes to this project will be documented in this file.  The format
 ## [Unreleased]
 
 ### Add
-* Added a new subcommand `get-era-summary` which optionally takes a block identifier and returns an era summary from a Casper network
+* Added a new subcommand `get-era-summary` which optionally takes a block identifier and returns an era summary from a Casper network.
 
 ### Changed
 * Update dependencies.